Save 'chronicle' objects where the value is a data.frame to disk

Save 'chronicle' objects where the value is a data.frame to disk

Usage

write_chronicle(.c, path, row.names = FALSE, sep = ",", ...)

Arguments

.c

A chronicle object.

path

Path to the file. The extension (one of '.csv' or '.xlsx') determines the format of the file.

row.names

Should row names be saved? Defaults to FALSE.

sep

The separator for '.csv' files. Defaults to ",".

...

Further arguments passed down to read.table() or openxlsx::write.xlsx().

Value

No return value, called for side effects (writing data to disk).

Details

'chronicle' objects where the "$value" is a data.frame can be saved with their log in either the '.csv' or '.xlsx' format to disk. In case the data is saved in the '.csv' format, the first n lines will contain the log. The number of lines that should be skipped to read in the data are indicated. In case the data gets saved in the '.xlsx' format, the data is written on the first sheet, and the log on the second sheet. read.table() with row.names = FALSE and sep = "," is used to save to the '.csv' format, and openxlsx::write.xlsx() is used to save to the '.xlsx' format.